Add ONSASConfig
